% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilebased-package.R
\docType{package}
\name{tilebased-package}
\alias{tilebased}
\alias{tilebased-package}
\title{tilebased: Traverse a Tile-Based Map in an R Graphics Window}
\description{
Use your keyboard to move a player sprite around the map.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
