% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.R
\name{play}
\alias{play}
\title{Begin a Game}
\usage{
play(height = 20L, width = 32L)
}
\arguments{
\item{height}{Integer. Number of tiles high for the map.}

\item{width}{Integer. Number of tiles wide for the map.}
}
\value{
Nothing. Draws to screen.
}
\description{
Begin a Game
}
\examples{
\dontrun{play()}
}
